Fix circular import between iptopo.py and ospf.py
